# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wallapop_API",
        "description": "API DE WALLAPOP",
        "contact": {
            "email": "daw011.2022@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "register a new user",
                "description": "Register by email, password and name",
                "operationId": "register",
                "requestBody": {
                    "description": "Register user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Juancito Perez"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login user",
                "description": "Login by email, password",
                "operationId": "Login",
                "requestBody": {
                    "description": "Login user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user1@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address or password. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout user",
                "description": "Logout User",
                "operationId": "Logout",
                "requestBody": {
                    "description": "Logout",
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Logout OK"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "description": "Muestra todos los posts paginados",
                "operationId": "show",
                "requestBody": {
                    "description": "Muestra todos los posts paginados",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Todos pueden ver los posts"
                    }
                }
            }
        },
        "/posts/addpost": {
            "post": {
                "tags": [
                    "posts"
                ],
                "description": "Permite crear un post nuevo",
                "operationId": "createPost",
                "requestBody": {
                    "description": "Create a new post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "category",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "Coche en venta"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Coche usado en buen estado"
                                    },
                                    "price": {
                                        "type": "string",
                                        "format": "decimal(10,2)",
                                        "example": "12,00"
                                    },
                                    "category": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Debe ser un numero"
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "file",
                                        "example": "image.jpg/png"
                                    },
                                    "persistent": {
                                        "type": "boolean",
                                        "example": "true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post creado correctamente"
                    }
                }
            }
        }
    }
}